# Optimize spring bones computations