--- a/packages/three-vrm-springbone/src/VRMSpringBoneColliderShape.ts
+++ b/packages/three-vrm-springbone/src/VRMSpringBoneColliderShape.ts
@@ -10,6 +10,11 @@ export abstract class VRMSpringBoneColliderShape {
   public abstract get type(): string;
 
   /**
+   * The offset to the shape.
+   */
+  public offset?: THREE.Vector3;
+
+  /**
    * Calculate a distance and a direction from the collider to a target object.
    * It's hit if the distance is negative.
    * The direction will be contained in the given target vector.