@0b5vr 2024-11-27
I'm kinda negative about the concept of the third commit. This change would prevent users from setting `offset` and `tail` after initializing, which is a breaking change in the first place.
We have yet to provide many editing features so far in three-vrm, but I still want to be conservative about making this change.

The save on the `length()` calculation looks great. Does this alone provide a performance improvement?