@mrxz 2024-11-27
> I'm kinda negative about the concept of the third commit. This change would prevent users from setting offset and tail after initializing, which is a breaking change in the first place.

I agree, the commit indeed makes changing the colliders after initialization difficult. This isn't a problem when just loading an avatar and using it, but users might want to implement editing features.

I've changed it so that the `colliderMatrix` is computed in the `updateWorldMatrix` method of `VRMSpringBoneCollider` instead. This way the `offset` and `tail` can still be changed and the collider matrix is still only computed once per update. The `VRMSpringBoneManager` guarantees that `updateWorldMatrix` is called for the colliders before updating the spring bones. Let me know what you think.